Update README.md
This Python script implements a simple to-do list application using object-oriented programming concepts.

ToDoList: This class represents a to-do list. It has methods to add a task, remove a task, and view all tasks.

init: Constructor method that initializes an empty list to store tasks. add_task: Method to add a task to the to-do list. remove_task: Method to remove a task from the to-do list. view_tasks: Method to view all tasks in the to-do list. The main block is responsible for running the to-do list application. It displays a menu with options to add a task, remove a task, view tasks, or quit the program.

When you run this script, it will prompt you with a menu to perform various operations on the to-do list such as adding tasks, removing tasks, viewing tasks, or quitting the program.